# openapi.yaml
openapi: 3.0.3
info:
  title: WarrantyVault API
  description: |
    REST API for native mobile clients (iOS / Android) of WarrantyVault.
    Source-of-truth for client codegen:
      - iOS:     swift-openapi-generator
      - Android: openapi-generator-cli (kotlin)
    The web app continues to use Next.js Server Actions — those are NOT
    documented here.

    Auth: bearer token from POST /api/v1/auth/login (or /register). Send as
    `Authorization: Bearer <token>` on every authenticated request.

    Ngôn ngữ: mọi chữ trong `message` / `fieldErrors` được server dịch sẵn. Thứ
    tự quyết định: `?lang=vi|en` → header `Accept-Language` → `User.locale` (đã
    lưu) → `en`. Ngôn ngữ không hỗ trợ rơi về mặc định, **không** phải lỗi 400.
    Xem `components.parameters.Lang`.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://warrantyvault.example.com
    description: Production (set when deployed)
...
security:
  - bearerAuth: []
paths:
  /api/v1/auth/register:
    parameters: [{ $ref: '#/components/parameters/Lang' }]
    post:
      tags: [auth]
      security: []
      summary: Create account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterInput' }
      responses:
        '201': { $ref: '#/components/responses/AuthSuccess' }
        '200':
          description: Email already taken (returns generic message; do not infer)
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean }, message: { type: string } }
        '400': { $ref: '#/components/responses/BadInput' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/auth/login:
    parameters: [{ $ref: '#/components/parameters/Lang' }]
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginInput' }
      responses:
        '200': { $ref: '#/components/responses/AuthSuccess' }
        '400': { $ref: '#/components/responses/BadInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/auth/logout:
    parameters: [{ $ref: '#/components/parameters/Lang' }]
    post:
      tags: [auth]
      summary: Revoke the current session token
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/forgot:
    parameters: [{ $ref: '#/components/parameters/Lang' }]
    post:
      tags: [auth]
      security: []
      summary: Request a password-reset link by email
      description: |
        Always returns `{ ok: true }` regardless of whether the email is
        registered (avoids user enumeration). Sends an email containing a
        time-limited reset link to the address if it exists.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForgotRequest' }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/BadInput' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/auth/me:
    parameters: [{ $ref: '#/components/parameters/Lang' }]
    get:
      tags: [auth]
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties: { user: { $ref: '#/components/schemas/User' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [auth]
      summary: Cập nhật hồ sơ — tên hiển thị và ngôn ngữ
      description: |
        Partial profile update. Two fields are accepted: `displayName` and
        `locale`. The account email **cannot be changed here** — it has its own
        two-step flow (`POST /api/v1/auth/change-email`). A body carrying
        `email` / `newEmail` is rejected with 400 and a `fieldErrors` entry
        saying so, rather than silently ignored.

        Both accepted fields are **tri-state** and independent: an ABSENT key
        leaves the stored value alone, `null` (or `""`) CLEARS it, and a value
        sets it. So `{"locale": "vi"}` on its own is a valid body, and a client
        that only knows about `displayName` cannot wipe a stored language by
        editing a name.

        A body with NEITHER key is 400, so an empty or mistyped body still fails
        loudly instead of being a 200 that changed nothing.

        Response mirrors `GET /api/v1/auth/me` (`{ "user": ... }`) plus a
        Vietnamese `message`, so a client can replace its cached user object
        directly from the PATCH response. Rate-limited per user like every other
        authenticated write (60 req/60s).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProfileInput' }
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  message: { type: string }
        '400': { $ref: '#/components/responses/BadInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
    delete:
      tags: [auth]
      summary: Permanently delete the authenticated user's account
      description: |
        Cascades the user's owned rows (Device / Subscription / Wishlist /
        Session / PasswordReset) in a single statement and best-effort
        removes the on-disk encrypted attachments. Requires the current
        password in the body for confirmation.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeleteAccountRequest' }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/BadInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/auth/reset-password:
    parameters: [{ $ref: '#/components/parameters/Lang' }]
    post:
      tags: [auth]
      security: []
      summary: Confirm a password-reset link by setting a new password
      description: |
        Takes the raw token from the `/forgot` link, finds the PasswordReset
        row by `sha256(token)`, updates the user's password hash, marks every
        outstanding reset row used, and revokes ALL of the user's Sessions
        (the password change kicks all devices off). Transactional.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
      responses:
        '200':
          description: Password reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  message: { type: string }
        '400': { $ref: '#/components/responses/BadInput' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/auth/change-password:
    parameters: [{ $ref: '#/components/parameters/Lang' }]
    post:
      tags: [auth]
      summary: Change the authenticated user's password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  message: { type: string }
        '400': { $ref: '#/components/responses/BadInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/auth/change-email:
    parameters: [{ $ref: '#/components/parameters/Lang' }]
    post:
      tags: [auth]
      summary: Yêu cầu đổi email — bước 1/2 (gửi token xác nhận tới địa chỉ mới)
      description: |
        Bước đầu của luồng đổi email. Body gồm **địa chỉ mới** và **mật khẩu hiện
        tại** (mật khẩu để một bearer token bị đánh cắp không thể tự chuyển tài
        khoản sang địa chỉ khác).

        Endpoint này **không** sửa `User.email`. Địa chỉ cũ vẫn đăng nhập / nhận
        thông báo bình thường cho tới khi bước 2 xác nhận thành công. Server tạo
        một token dùng-một-lần, hết hạn sau **30 phút** (cùng hạ tầng
        `PasswordReset` với luồng quên mật khẩu, cột `pendingEmail`) và gửi email
        tới **địa chỉ mới** — nhận được email ở đó chính là bằng chứng sở hữu địa
        chỉ. Email chứa cả link (`{APP_URL}/confirm-email/{token}`) và mã token
        thô để client native gọi thẳng bước 2.

        **Không tiết lộ địa chỉ đã tồn tại:** nếu địa chỉ mới đã thuộc một tài
        khoản khác, phản hồi vẫn là 200 với *cùng một* `message` trung tính như
        khi thành công, và **không** có email nào được gửi (giống cách
        `POST /api/v1/auth/register` xử lý email đã tồn tại). Chỉ khi địa chỉ mới
        trùng với chính email hiện tại của bạn thì mới trả 400.

        Yêu cầu bearer token. Rate limit theo user.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangeEmailRequest' }
      responses:
        '200':
          description: |
            Đã ghi nhận yêu cầu (hoặc địa chỉ đã được dùng — hai trường hợp trả
            cùng một thông báo trung tính).
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  message: { type: string }
        '400': { $ref: '#/components/responses/BadInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/auth/confirm-email-change:
    parameters: [{ $ref: '#/components/parameters/Lang' }]
    post:
      tags: [auth]
      summary: Xác nhận đổi email — bước 2/2 (tiêu thụ token)
      description: |
        Bước cuối của luồng đổi email. Body chỉ có `token` (token thô trong email
        gửi tới địa chỉ mới). **Không yêu cầu bearer token**: token là chứng thực,
        và link thường được mở trên thiết bị khác.

        Khi thành công, trong một transaction: cập nhật `User.email` sang địa chỉ
        mới, đánh dấu token đã dùng **và** huỷ mọi token còn treo khác của user
        (cả token đặt lại mật khẩu lẫn token đổi email), rồi **thu hồi toàn bộ
        session** — giống hệt luồng đặt lại mật khẩu. Vì vậy client phải đăng nhập
        lại bằng địa chỉ mới; các thiết bị khác cũng bị đăng xuất.

        Token dùng một lần và hết hạn sau 30 phút. Token đã dùng, token hết hạn,
        token không tồn tại, **và token đặt lại mật khẩu** đều trả 400
        `invalid_email_change_token` (hai loại token được phân biệt bằng cột
        `pendingEmail`, migration 0009 — token đổi email không thể dùng để đặt lại
        mật khẩu và ngược lại). Nếu địa chỉ mới bị người khác đăng ký trong lúc
        chờ xác nhận: 400 `email_in_use`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, description: 'Token thô trong email xác nhận.' }
      responses:
        '200':
          description: Đã đổi email (mọi session đã bị thu hồi)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  message: { type: string }
        '400': { $ref: '#/components/responses/BadInput' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /api/v1/push/test:
    parameters: [{ $ref: '#/components/parameters/Lang' }]
    post:
      tags: [push]
      summary: Send a sample notification to every registered subscription
      description: |
        Dispatches a test notification through the push fanout (web / APNs /
        FCM) to every PushSubscription the caller owns. Subscriptions the
        upstream gateway reports as gone (404/410, FCM UNREGISTERED) are
        deleted as a side effect, matching the cron's gone-detection.

        The notification BODY is translated like the cron's are, so `?lang=`
        changes what lands on the device — this is the one push whose language a
        client can set per request rather than through `User.locale`.
      responses:
        '200':
          description: Send counts
          content:
            application/json:
              schema:
                type: object
                required: [sent, failed]
                properties:
                  sent:
                    type: integer
                    minimum: 0
                    description: Subscriptions the dispatcher accepted
                  failed:
                    type: integer
                    minimum: 0
                    description: Subscriptions that errored (including gone rows that were deleted)
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
...
    Lang:
      name: lang
      in: query
      required: false
      schema: { type: string, enum: [vi, en] }
      description: |
        Ngôn ngữ của phần chữ trong response (`message`, `fieldErrors`). Thắng cả
        header `Accept-Language` lẫn lựa chọn đã lưu. Giá trị khác `vi`/`en` bị
        bỏ qua (rơi xuống mức kế tiếp), **không** phải lỗi 400.
  responses:
    Ok:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties: { ok: { type: boolean } }
    AuthSuccess:
      description: Authenticated, returns access token + user
      content:
        application/json:
          schema:
            type: object
            required: [accessToken, expiresAt, user]
            properties:
              accessToken: { type: string }
              expiresAt: { type: string, format: date-time }
              user: { $ref: '#/components/schemas/User' }
    BadInput:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: Missing / invalid / expired bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Resource not found (or owned by another user)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string, description: 'Stable code, e.g. unauthorized, not_found, bad_input, rate_limited, limit_reached' }
        message: { type: string, description: 'Human-readable Vietnamese message' }
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
...
    User:
      type: object
      required: [id, email]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string, nullable: true }
        aiOptIn: { type: boolean, description: 'Người dùng đã bật tính năng quét hoá đơn AI (mặc định false)' }
        locale:
          type: string
          nullable: true
          enum: [en, vi]
          description: |
            Ngôn ngữ người dùng đã chọn (`User.locale`, migration 0014). `null`
            nghĩa là **chưa chọn** — KHÔNG phải "tiếng Anh": request vẫn quyết
            định theo thứ tự `?lang=` → `Accept-Language` → cột này → `en`.

            Field này **vắng mặt** trong response khi người dùng chưa từng chọn
            ngôn ngữ, để response của client cũ không đổi một byte nào. Chỉ nhận
            hai mã hai chữ cái; gửi `vi-VN` thì server lưu `vi`.
    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 200 }
        name: { type: string, maxLength: 80, nullable: true }
        deviceLabel: { type: string, maxLength: 80, nullable: true }
        platform: { type: string, enum: [ios, android, web], nullable: true }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
        deviceLabel: { type: string, maxLength: 80, nullable: true }
        platform: { type: string, enum: [ios, android, web], nullable: true }
    ForgotRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }

    # Body of PATCH /api/v1/auth/me. Email change is NOT part of this contract.
    UpdateProfileInput:
      type: object
      description: |
        Chỉ `displayName` được chấp nhận. Đổi email **không** gửi ở đây — có luồng
        riêng hai bước (`POST /api/v1/auth/change-email` rồi
        `POST /api/v1/auth/confirm-email-change`). Gửi thêm `email`/`newEmail` vào
        endpoint này sẽ trả 400 kèm `fieldErrors` chỉ đúng hai endpoint đó.
      properties:
        displayName:
          type: string
          nullable: true
          maxLength: 80
          description: |
            Tên hiển thị mới. `null`, chuỗi rỗng hoặc chỉ khoảng trắng ⇒ XOÁ tên
            (`User.name` trở thành null). Giới hạn 80 **byte** UTF-8, đúng bằng
            giới hạn của bước đăng ký (tên tiếng Việt có dấu tốn nhiều byte hơn
            số ký tự). Giá trị được trim trước khi lưu.

            **Vắng mặt ⇒ giữ nguyên** tên đang có (khác hẳn `null`). Đây là quy
            ước ba trạng thái mà `locale` cũng theo, và là lý do hai field này
            độc lập nhau: client cũ chỉ gửi `displayName` vẫn không thể vô tình
            xoá ngôn ngữ đã lưu.
        locale:
          type: string
          nullable: true
          enum: [en, vi]
          description: |
            Ngôn ngữ của chính người dùng. Ba trạng thái:

            * **vắng mặt** ⇒ giữ nguyên giá trị đã lưu;
            * `null` hoặc `""` ⇒ XOÁ (về `null`), tức quay lại để request quyết
              định theo `Accept-Language`;
            * `"en"` / `"vi"` (chấp nhận cả `vi-VN`, `en-US` — lưu mã gốc) ⇒ đặt.

            Giá trị không hỗ trợ (ví dụ `"fr"`) là **400** kèm
            `fieldErrors.locale`, KHÁC với một `Accept-Language: fr` (cái đó chỉ
            rơi về mặc định): đây là một write hỏng, còn kia là một sở thích
            chúng tôi không có.

            Gửi body **rỗng** (không có cả `displayName` lẫn `locale`) là 400.

            Ảnh hưởng: ngôn ngữ của `message`/`fieldErrors` trong mọi response
            sau đó (khi request không nói gì), và ngôn ngữ của **push
            notification** — cron không có request nào để hỏi, nên cột này là
            tín hiệu duy nhất. Xem `components.parameters.Lang`.

    # Body of POST /api/v1/auth/change-email (step 1 of 2).
    ChangeEmailRequest:
      type: object
      required: [newEmail, currentPassword]
      properties:
        newEmail:
          type: string
          format: email
          description: |
            Địa chỉ muốn chuyển sang. Được trim + hạ chữ thường trước khi lưu.
            Nếu đã thuộc tài khoản khác, phản hồi vẫn là 200 trung tính và không
            có email nào được gửi (không tiết lộ địa chỉ tồn tại hay không).
        currentPassword:
          type: string
          minLength: 1
          description: 'Mật khẩu hiện tại của tài khoản. Sai ⇒ 400 với `fieldErrors.currentPassword`.'
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword, confirmPassword]
      properties:
        currentPassword: { type: string, minLength: 1 }
        newPassword: { type: string, minLength: 8, maxLength: 200 }
        confirmPassword: { type: string, minLength: 1 }
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token: { type: string, description: 'Raw token from the /forgot reset link' }
        newPassword: { type: string, minLength: 8, maxLength: 200 }
    DeleteAccountRequest:
      type: object
      required: [password]
      properties:
        password: { type: string, minLength: 1, description: 'Current password — required for confirmation' }
...